loadforge/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
redis==5.0.4
locust==2.28.0
python-dotenv==1.0.1
